CTest-integrated benchmarks: 102 tests pass on ESP32-P4
Benchmarks run as doctest test cases (test_bench.cpp) tagged with
test_suite("bench"), integrated into both desktop CTest and ESP32-P4
CTest infrastructure. Same methodology: validate → warm-up → 3 measured
passes → median fps.

ESP32-P4 benchmark results (360 MHz, -O2, PSRAM 200 MHz):
  Baseline CAVLC (short):  50 frames, 2911 ms median, 17.2 fps (CoV 0.07%)
  High CABAC:             195 frames, 4020 ms median, 48.5 fps * (CoV 0.03%)
  Flat black:             250 frames, 22547 ms median, 11.1 fps (CoV 0.01%)

  * CABAC fps inflated — neutral context init, full m,n tables pending

  Baseline 17.2 fps exceeds NestNinja S2 reference (16.71 fps) by 3%.

Desktop benchmark results (x86 Release, MSVC):
  Baseline CAVLC (short): ~1042 fps
  High CABAC:             ~5008 fps *
  Flat black:              ~703 fps

Changes:
- test_bench.cpp: Benchmark test cases with validate/warmup/measure
- bench_decode.cpp: Updated to use decode_timing.hpp timer
- decode_timing.hpp: Fixed printf format for RISC-V (uint32_t = long)
- CMakeLists.txt (tests): Added test_bench.cpp + self-include path
- CMakeLists.txt (unit_tests): Extended timeouts (480s test, 600s CTest)
- flash_and_test.cmake: Removed hardcoded 180s TIMEOUT from
  execute_process — let Python script and CTest handle timeouts
- test_runner.cpp: Include test_bench.cpp for ESP32

102 test cases (99 unit + 3 bench), all passing on ESP32-P4.
101 test cases on desktop, all passing.

+++ components/sub0h264/src/decode_timing.hpp
@@ -67,7 +67,7 @@ struct StreamTiming
             ? (100.0 * totalReconUs / totalFrameUs) : 0.0;
         double overheadPct = 100.0 - parsePct - reconPct;
 
-        std::printf("  Parse/Recon split (%u frames):\n", frameCount);
+        std::printf("  Parse/Recon split (%lu frames):\n", (unsigned long)frameCount);
         std::printf("    Parse:    %8lld us  (%5.1f%%)\n",
                     (long long)totalParseUs, parsePct);
         std::printf("    Recon:    %8lld us  (%5.1f%%)\n",
